refactor player customisation functions into own file
