+++ samples/golf/golf.vcxproj.filters
@@ -189,6 +189,9 @@
     <ClCompile Include="src\golf\spooky2.cpp">
       <Filter>Source Files\golf\client</Filter>
     </ClCompile>
+    <ClCompile Include="src\golf\MenuCustomisation.cpp">
+      <Filter>Source Files\golf\client\states</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\ErrorCheck.hpp">
