+++ samples/golf/src/golf/MenuState.cpp
@@ -29,7 +29,6 @@ source distribution.
 
 #include "MenuState.hpp"
 #include "MenuSoundDirector.hpp"
-#include "SharedStateData.hpp"
 #include "PacketIDs.hpp"
 #include "MenuConsts.hpp"
 #include "Utility.hpp"
@@ -716,151 +715,6 @@ void MenuState::createScene()
     createBallScene();    
 
     createUI();
-}
-
-void MenuState::createBallScene()
-{
-    static constexpr float RootPoint = 100.f;
-    static constexpr float BallSpacing = 0.09f;
-
-    auto ballTexCallback = [&](cro::Camera&)
-    {
-        auto vpSize = calcVPSize().y;
-        auto windowSize = static_cast<float>(cro::App::getWindow().getSize().y);
-
-        float windowScale = std::floor(windowSize / vpSize);
-        float scale = m_sharedData.pixelScale ? windowScale : 1.f;
-        auto size = BallPreviewSize * static_cast<std::uint32_t>((windowScale + 1.f) - scale);
-        m_ballTexture.create(size, size);
-    };
-
-    m_ballCam = m_backgroundScene.createEntity();
-    m_ballCam.addComponent<cro::Transform>().setPosition({ RootPoint, 0.045f, 0.095f });
-    m_ballCam.getComponent<cro::Transform>().rotate(cro::Transform::X_AXIS, -0.03f);
-    m_ballCam.addComponent<cro::Camera>().setPerspective(1.f, 1.f, 0.001f, 2.f);
-    m_ballCam.getComponent<cro::Camera>().resizeCallback = ballTexCallback;
-    m_ballCam.addComponent<cro::Callback>().active = true;
-    m_ballCam.getComponent<cro::Callback>().setUserData<std::int32_t>(0);
-    m_ballCam.getComponent<cro::Callback>().function =
-        [](cro::Entity e, float dt)
-    {
-        auto id = e.getComponent<cro::Callback>().getUserData<std::int32_t>();
-        float target = RootPoint + (BallSpacing * id);
-
-        auto pos = e.getComponent<cro::Transform>().getPosition();
-        auto diff = target - pos.x;
-        pos.x += diff * (dt * 10.f);
-
-        e.getComponent<cro::Transform>().setPosition(pos);
-    };
-
-    ballTexCallback(m_ballCam.getComponent<cro::Camera>());
-
-
-    auto ballFiles = cro::FileSystem::listFiles(cro::FileSystem::getResourcePath() + "assets/golf/balls");
-    if (ballFiles.empty())
-    {
-        LogE << "No ball files were found" << std::endl;
-    }
-
-    m_sharedData.ballModels.clear();
-    
-    for (const auto& file : ballFiles)
-    {
-        cro::ConfigFile cfg;
-        if (cro::FileSystem::getFileExtension(file) == ".ball"
-            && cfg.loadFromFile("assets/golf/balls/" + file))
-        {
-            std::uint32_t uid = SpookyHash::Hash32(file.data(), file.size(), 0);
-            std::string modelPath;
-            cro::Colour colour = cro::Colour::White;
-
-            const auto& props = cfg.getProperties();
-            for (const auto& p : props)
-            {
-                const auto& name = p.getName();
-                if (name == "model")
-                {
-                    modelPath = p.getValue<std::string>();
-                }
-                /*else if (name == "uid")
-                {
-                    uid = p.getValue<std::int32_t>();
-                }*/
-                else if (name == "tint")
-                {
-                    colour = p.getValue<cro::Colour>();
-                }
-            }
-
-            if (/*uid > -1
-                &&*/ (!modelPath.empty() && cro::FileSystem::fileExists(cro::FileSystem::getResourcePath() + modelPath)))
-            {
-                auto ball = std::find_if(m_sharedData.ballModels.begin(), m_sharedData.ballModels.end(),
-                    [uid](const SharedStateData::BallInfo& ballPair)
-                    {
-                        return ballPair.uid == uid;
-                    });
-
-                if (ball == m_sharedData.ballModels.end())
-                {
-                    m_sharedData.ballModels.emplace_back(colour, uid, modelPath);
-                }
-                else
-                {
-                    LogE << file << ": a ball already exists with UID " << uid << std::endl;
-                }
-            }
-        }
-    }
-
-    cro::ModelDefinition ballDef(m_resources);
-
-    cro::ModelDefinition shadowDef(m_resources);
-    auto shadow = shadowDef.loadFromFile("assets/golf/models/ball_shadow.cmt");
-
-    for (auto i = 0u; i < m_sharedData.ballModels.size(); ++i)
-    {
-        if (ballDef.loadFromFile(m_sharedData.ballModels[i].modelPath))
-        {
-            auto entity = m_backgroundScene.createEntity();
-            entity.addComponent<cro::Transform>().setPosition({ (i * BallSpacing) + RootPoint, 0.f, 0.f });
-            ballDef.createModel(entity);
-            entity.getComponent<cro::Model>().setMaterial(0, m_resources.materials.get(m_materialIDs[MaterialID::Cel]));
-
-            entity.addComponent<cro::Callback>().active = true;
-            entity.getComponent<cro::Callback>().function =
-                [](cro::Entity e, float dt)
-            {
-                e.getComponent<cro::Transform>().rotate(cro::Transform::Y_AXIS, /*0.3f **/ dt);
-            };
-
-            if (shadow)
-            {
-                auto ballEnt = entity;
-                entity = m_backgroundScene.createEntity();
-                entity.addComponent<cro::Transform>();
-                shadowDef.createModel(entity);
-                ballEnt.getComponent<cro::Transform>().addChild(entity.getComponent<cro::Transform>());
-            }
-        }
-    }
-}
-
-std::int32_t MenuState::indexFromBallID(std::uint32_t ballID)
-{
-    auto ball = std::find_if(m_sharedData.ballModels.begin(), m_sharedData.ballModels.end(),
-        [ballID](const SharedStateData::BallInfo& ballPair)
-        {
-            return ballPair.uid == ballID;
-        });
-
-    if (ball != m_sharedData.ballModels.end())
-    {
-        return static_cast<std::int32_t>(std::distance(m_sharedData.ballModels.begin(), ball));
-    }
-
-    return -1;
 }
 
 void MenuState::handleNetEvent(const cro::NetEvent& evt)
